feat: visualize type on click
