+++ .gitignore
@@ -1,3 +1,4 @@
 /venv
 /__pycache__
-*.html
+QEP.html
+output.log
